requirements.txt
base58
fastecdsa
protobuf
pycrypto
